% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Anolis}
\alias{Anolis}
\title{A SNP data for two Anolis species; 25 SNPs used to test for signals of differential introgression.}
\format{
A data frame that includes sample labels, population assignments, and genotype information for 73 individuals.
}
\source{
Farleigh et al. (in prep)
}
\usage{
data(Anolis)
}
\description{
SNP data for 73 individuals from a study used to infer differential introgression between
Anolis pulchellus and A. krugi. An example data file demonstrating how data should be formatted
for use in HybridFindR.
}
\examples{
data(Anolis)
str(Anolis)
}
\references{
Farleigh et al. (in prep)
}
\keyword{datasets}
