% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Differential_introgression.R
\name{Differential_introgression}
\alias{Differential_introgression}
\title{Peroform differential introgression analysis}
\usage{
Differential_introgression(
  Data,
  H.id,
  S0.id,
  S1.id,
  n.ind,
  permutations,
  ploidy
)
}
\arguments{
\item{Data}{Object produced by the PrepareData function}

\item{H.id}{Character, the label for the hybrid population.}

\item{S0.id}{Character, the label for the S0 parental population.}

\item{S1.id}{Character, the label for the S1 parental population.}

\item{n.ind}{Numeric, the number of individuals in your data file.}

\item{permutations}{Numeric, number of permutations to use in significance testing.}

\item{ploidy}{Numeric, ploidy of markers in the Data object.}
}
\value{
A p-value for each SNP representing the proportion of permutations that possessed a greater proportion of S1 alleles in permuted hybrids than the empirical data.
}
\description{
Peroform differential introgression analysis
}
\examples{
\dontrun{ DI_test <- Differential_introgression(Data, H.id = 'H', S0.id = 'P',
S1.id = 'K', n.ind = 73, permutations = 1000, ploidy = 2)}
}
