% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loci_stats.R
\name{Loci_stats}
\alias{Loci_stats}
\title{Analyze each locus to determine which individuals are homozygous and heterozygous.}
\usage{
Loci_stats(Data)
}
\arguments{
\item{Data}{object produced by the PrepareData function.}
}
\value{
A list indicating which individuals are homozygous for the S1 and S0 alleles at each locus as well as the individuals that are heterozygous at each locus.
}
\description{
Analyze each locus to determine which individuals are homozygous and heterozygous.
}
\examples{
\dontrun{ Loc_stats <- Loci_stats(Data)}
}
