% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Missingness.R
\name{Missingness}
\alias{Missingness}
\title{Analyze the amount of missing data in your dataset}
\usage{
Missingness(Data)
}
\arguments{
\item{Data}{Object produced by the PrepareData function}
}
\value{
A list containing the average missingness per individual and per locus as well as a barplot for each.
}
\description{
Analyze the amount of missing data in your dataset
}
\examples{
\dontrun{ Missing <- Missingness(Data)}
}
