% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotprops.R
\name{Plotprops}
\alias{Plotprops}
\title{Plot the proportion of parental alleles in hybrid individuals}
\usage{
Plotprops(Data, SNP_names, ploidy, S1_color, S0_color)
}
\arguments{
\item{Data}{Object produced by the PrepareData function}

\item{SNP_names}{Character or character vector, containing the names of SNPs that you want to plot.}

\item{ploidy}{Numeric, ploidy of markers in the Data object.}

\item{S1_color}{Character, desired color for the S1 parental alleles.}

\item{S0_color}{Character, desired color for the S0 parental alleles.}
}
\value{
A plot displaying the proportion of parental alleles in hybrids at each SNP provided to the SNP_names parameter.
}
\description{
Plot the proportion of parental alleles in hybrid individuals
}
\examples{
\dontrun{ Plotprops(SNP_names = candidate_snps, ploidy = 2,
S1_color = 'red', S0_color = 'blue')}
}
