% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{PrepareData}
\alias{PrepareData}
\title{Prepare a structure file (or similar data table) for differential introgression analysis using gghybrid.}
\usage{
PrepareData(Data, numprecol, missingval, onerow, numinds, S0.id, S1.id)
}
\arguments{
\item{Data}{Character string defining the structure or data table to read in.}

\item{numprecol}{Numeric, the number of columns preceding the genetic data.}

\item{missingval}{Numeric or character, The value for indicating missing data.}

\item{onerow}{Numeric, (0 or 1). Whether or not there is a single row per individual. Currently, HybridFindR only supports 0.}

\item{numinds}{Numeric, the number of individuals in your data file.}

\item{S0.id}{Character, the label for the S0 parental population.}

\item{S1.id}{Character, the label for the S1 parental population.}
}
\value{
Returns a genotype table based on the number of copies of the S1 allele in each individual at each marker.
}
\description{
Prepare a structure file (or similar data table) for differential introgression analysis using gghybrid.
}
\examples{
\dontrun{
Data <- PrepareData('mystructurefile.csv', numprecol = 2, missingval = 'NA',
onerow = 0, numinds = 73, S0.id = 'P', S1.id = 'K')}
}
\references{
Bailey R. (2020) gghybrid: R package for evolutionary analysis of hybrids and hybrid zones.10.5281/zenodo.3676499
}
